pipeline_export/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
easyocr>=1.7.0
html5lib==1.1
idna==3.10
numerize==0.12
openai>=1.0.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
supabase>=2.0.0
ua-generator==2.0.0
urllib3==2.3.0
webencodings==0.5.1
boto3>=1.28.0
